# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Application API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the generated application",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "https://your-production-url.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "security": [],
  "paths": {
...
    "/api/seo/{id}": {
      "get": {
        "summary": "Get SEO campaign by ID",
        "tags": [
          "SEO"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SEO campaign details"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/seo": {
      "get": {
        "summary": "Get all SEO campaigns",
        "tags": [
          "SEO"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of SEO campaigns"
          }
        }
      },
      "post": {
        "summary": "Create new SEO campaign",
        "tags": [
          "SEO"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "SEO campaign created"
          }
        }
      }
    },
    "/api/websites/{id}": {
      "get": {
        "summary": "Get website by ID",
        "tags": [
          "Websites"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Website details"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/websites": {
      "get": {
        "summary": "Get all websites",
        "tags": [
          "Websites"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of websites"
          }
        }
      },
      "post": {
        "summary": "Create new website",
        "tags": [
          "Websites"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Website created"
          }
        }
      }
    }
  },
...
}